Extract common methods from TestCaseExtractor objects to parent trait
